Updated heaps of stuff to use new temperature sensor. Temperature is now a int32 throughout.

+++ state-machine.h
@@ -19,7 +19,7 @@ struct Inputs {
     uint8_t waiting_on_inputs;
     
     /* These values contain the input metrics for the state machine */
-    uint16_t temperature; // degrees celcius (Cx100)
+    int32_t temperature; // degrees celcius (Cx100)
     uint16_t level; // height (mm)
     uint16_t volume; // Litres
     
@@ -57,11 +57,11 @@ struct Settings {
     uint16_t level_min;         // The minimum allowable level of water in the tank
     uint16_t level_fill;        // The level to fill the tank to\
     /* Temperature settings */
-    uint16_t temperature_settemp;
-    uint16_t temperature_set_1;  // The temperature to which the system will automatically be heated
-    uint16_t temperature_set_2;
-    uint16_t temperature_max;    // Maximum temperature to maintain
-    uint16_t temperature_min;    // Minimum temperature to maintain
+    int32_t temperature_settemp;
+    int32_t temperature_set_1;  // The temperature to which the system will automatically be heated
+    int32_t temperature_set_2;
+    int32_t temperature_max;    // Maximum temperature to maintain
+    int32_t temperature_min;    // Minimum temperature to maintain
     /* Smarts */
     uint16_t daily_heat_potential;      // How much energy to expect in a day
     uint32_t midsun;            // Time at which the sun was/is at its peak
